Use stronger Anti-CSWSH mechanism
I went with a variation of the "encrypted token pattern". (The variation
is that we're still storing the token in a cookie as a mechansim for
JavaScript to get the cookie, taking care to only read the value out of
the query string.)

This will help protect against cross-domain cookie shennanigans, along
with who knows what else.

 The demo app doesn't have authentication so the user id is always null,
 but I wanted to show how it might be done.